Firebase crud basic functionality ended, nav pending

+++ frontendapp/src/Components/IssueManagement/Show.js
@@ -4,6 +4,9 @@ import firebase from "../../Firebase/Firebase";
 import { Link } from "react-router-dom";
 import Header from "../../layouts/Header";
 import { Grid, Paper } from "@material-ui/core";
+//Import Alertify
+import Swal from "sweetalert2/dist/sweetalert2.js";
+import "sweetalert2/src/sweetalert2.scss";
 
 export default class Show extends Component {
   constructor(props) {
@@ -33,18 +36,31 @@ export default class Show extends Component {
   }
 
   delete = () => {
-    firebase
-      .firestore()
-      .collection("issues")
-      .doc(this.state.key)
-      .delete()
-      .then(() => {
-        console.log("Document successfully deleted!");
-        this.props.history.push("/issues");
-      })
-      .catch(error => {
-        console.error("Error removing document: ", error);
-      });
+    Swal.fire({
+      title: "Delete this issue?",
+      text: "You won't be able to revert this, be wise!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then(result => {
+      if (result.value) {
+        firebase
+          .firestore()
+          .collection("issues")
+          .doc(this.state.key)
+          .delete()
+          .then(() => {
+            console.log("Issues successfully deleted!");
+            this.props.history.push("/issues");
+          })
+          .catch(error => {
+            console.error("Error removing document: ", error);
+          });
+        Swal.fire("Deleted!", "Your Issue has been deleted.", "success");
+      }
+    });
   };
 
   render() {
@@ -56,13 +72,12 @@ export default class Show extends Component {
             <Paper style={{ padding: 20, marginTop: 10, marginBottom: 10 }}>
               <div className="panel panel-default">
                 <div className="panel-heading">
-                  <h4>
-                    <Link to="/issues">RTC Issues List</Link>
-                  </h4>
                   <h3 className="panel-title">{this.state.issue.subject}</h3>
                 </div>
                 <div className="panel-body">
                   <dl>
+                  <dt>Case number:</dt>
+                    <dd>{this.state.issue.case_number}</dd>
                     <dt>PFI Tag used:</dt>
                     <dd>{this.state.issue.pfi_tag}</dd>
                     <dt>Product:</dt>
@@ -79,9 +94,16 @@ export default class Show extends Component {
                     Edit
                   </Link>
                   &nbsp;
-                  <button onClick={this.delete} className="btn btn-danger">
+                  <button
+                    style={{ marginRight: 5 }}
+                    onClick={this.delete}
+                    className="btn btn-danger"
+                  >
                     Delete
                   </button>
+                  <Link to="/issues" className="btn btn-primary">
+                    Go back Issues List
+                  </Link>
                 </div>
               </div>
             </Paper>
